+++ data_structures/linked_lists/test_linked_list.py
@@ -1,5 +1,4 @@
-from single_linked_list import LinkedList, ListNode
-
+from .linked_list import LinkedList
 import pytest
